refactor(api): move authenticateForSaml usecase in IAM context
