+++ api/src/identity-access-management/domain/usecases/index.js
@@ -13,6 +13,7 @@ import { config } from '../../../shared/config.js';
 import { eventBus } from '../../../shared/domain/events/index.js';
 import { cryptoService } from '../../../shared/domain/services/crypto-service.js';
 import { mailService } from '../../../shared/domain/services/mail-service.js';
+import * as obfuscationService from '../../../shared/domain/services/obfuscation-service.js';
 import { tokenService } from '../../../shared/domain/services/token-service.js';
 import * as userService from '../../../shared/domain/services/user-service.js';
 import * as passwordValidator from '../../../shared/domain/validators/password-validator.js';
@@ -92,6 +93,7 @@ const services = {
   authenticationSessionService,
   cryptoService,
   mailService,
+  obfuscationService,
   oidcAuthenticationServiceRegistry,
   passwordGeneratorService,
   pixAuthenticationService,
